+++ items/layouts.sh
@@ -37,7 +37,8 @@ static_content() {
     set -- \
         0.0 M Home "Back to Main menu  $nav_home" main.sh \
         2.5 M L "Border Lines" layouts_lines.sh \
-        3.3 M I "Border Indicators" layouts_indicators.sh
+        3.3 M I "Border Indicators" layouts_indicators.sh \
+        3.6 M S "Scroll Bars" layouts_scrollbars.sh
     menu_generate_part 2 "$@"
     $cfg_display_cmds && display_commands_toggle 3